+++ packages/gatsby-transformer-react-docgen/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.19](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.18...gatsby-transformer-react-docgen@5.0.19) (2019-10-14)
+
+**Note:** Version bump only for package gatsby-transformer-react-docgen
+
+## [5.0.18](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.17...gatsby-transformer-react-docgen@5.0.18) (2019-10-14)
+
+**Note:** Version bump only for package gatsby-transformer-react-docgen
+
+## [5.0.17](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.16...gatsby-transformer-react-docgen@5.0.17) (2019-10-09)
+
+**Note:** Version bump only for package gatsby-transformer-react-docgen
+
+## [5.0.16](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.15...gatsby-transformer-react-docgen@5.0.16) (2019-10-04)
+
+**Note:** Version bump only for package gatsby-transformer-react-docgen
+
+## [5.0.15](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.13...gatsby-transformer-react-docgen@5.0.15) (2019-09-26)
+
+**Note:** Version bump only for package gatsby-transformer-react-docgen
+
+## [5.0.14](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.13...gatsby-transformer-react-docgen@5.0.14) (2019-09-26)
+
+**Note:** Version bump only for package gatsby-transformer-react-docgen
+
+## [5.0.13](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.12...gatsby-transformer-react-docgen@5.0.13) (2019-09-20)
+
+**Note:** Version bump only for package gatsby-transformer-react-docgen
+
+## [5.0.12](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.11...gatsby-transformer-react-docgen@5.0.12) (2019-09-10)
+
+**Note:** Version bump only for package gatsby-transformer-react-docgen
+
+## [5.0.11](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.10...gatsby-transformer-react-docgen@5.0.11) (2019-09-09)
+
+**Note:** Version bump only for package gatsby-transformer-react-docgen
+
+## [5.0.10](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.9...gatsby-transformer-react-docgen@5.0.10) (2019-09-09)
+
+**Note:** Version bump only for package gatsby-transformer-react-docgen
+
 ## [5.0.9](https://github.com/gatsbyjs/gatsby/compare/gatsby-transformer-react-docgen@5.0.8...gatsby-transformer-react-docgen@5.0.9) (2019-09-01)
 
 ### Bug Fixes